add component and model bottom navigation bar
