+++ lib/Scenes/Page/main_page.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:timeboxing/Scenes/Page/HomePage/home_page.dart';
-import 'package:timeboxing/Shared/Widget/bottom_navigation_bar.dart';
+import 'package:timeboxing/Shared/Widget/Component/bottom_navigation_bar.dart';
 
 class MainPage extends StatelessWidget {
   const MainPage({super.key});
