+++ lib/Shared/Widget/Component/bottom_navigation_bar.dart
@@ -5,20 +5,7 @@ import 'package:timeboxing/Scenes/Page/HomePage/home_page.dart';
 import 'package:timeboxing/Scenes/Page/ProfilePage/profile_page.dart';
 import 'package:timeboxing/Shared/Extension/colors_style_extension.dart';
 import 'package:timeboxing/Shared/Extension/icons_style_extension.dart';
-
-class TabBarItem {
-  IconData iconData;
-  String title;
-  Widget page;
-  Color iconColor;
-
-  TabBarItem({
-    required this.iconData,
-    required this.title,
-    required this.page,
-    required this.iconColor,
-  });
-}
+import 'package:timeboxing/Shared/Widget/Model/tab_bar_item.dart';
 
 class TimeboxingBottomNavigationBar extends StatefulWidget {
   const TimeboxingBottomNavigationBar({super.key});
