+++ lib/Shared/Widget/Model/tab_bar_item.dart
@@ -1,0 +1,15 @@
+import 'package:flutter/material.dart';
+
+class TabBarItem {
+  IconData iconData;
+  String title;
+  Widget page;
+  Color iconColor;
+
+  TabBarItem({
+    required this.iconData,
+    required this.title,
+    required this.page,
+    required this.iconColor,
+  });
+}
